GPA species information

A species object defining the life history and function for green peach aphid (Myzus persicae). Can be used to run the endosymbiont model on GPA.

f_growth fitted using fit_custom with data from Whalon & Smilowitz 1979.
f_loss fitted using fit_sigmoid with data from Maling et al. 2008.
f_suscept fitted using fit_custom with data from Thackray et al. 2009.
f_inoc fitted using fit_quadratic with data from Thackray et al. 2009.
f_alate fitted using fit_custom with data from Thackray et al. 2009.

Usage

data(GPA)

Format

An object of class species; see species.

References

Maling, T., Diggle, A.J., Thackray, D.J., Siddique, K.H.M., Jones, R.A.C., An epidemiological model for externally sourced vector-borne viruses applied to Bean yellow mosaic virus in lupin crops in a Mediterranean-type environment. Phytopathology, 98, 1280–1290 (2008)

Thackray, D.J., Diggle, A.J., Jones, R.A.C., BYDV PREDICTOR: a simulation model to predict aphid arrival, epidemics of Barley yellow dwarf virus and yield loss in wheat crops in a Mediterranean-type environment. Plant Pathology, 58, 186–202 (2009)]

Whalon, M.E., Smilowitz, Z., Temperature-dependent model for predicting field populations of green peach aphid, Myzus persicae. The Canadian Entomologist, 111, 1025–1032 (1979)