An S4 class to store a pest's life history for the endosymbiont model

An S4 class to store a pest's life history for the endosymbiont model

Slots

species

The name of the pest

bg_loss

Background mortality (0-1)

alate_penalty

Relative fecundity penalty of alates compaered to apterae (0-1)

apterae_walk

Number of plants per day visited by apterae

alate_flight

Number of plants per day visited by alates

fun_dev_apt

Function describing temperature-dependent development of nymphs destined to become apterae

fun_dev_ala

Function describing temperature-dependent development of nymphs destined to become alates

fun_imi

Function describing number of immigrant alates

fun_emi

Function describing number of emigrants among newly metamorphosed alates

fun_temp_loss

Function describing daily mortality due to exceeding temperature thresholds

fun_rainfall_loss

Function describing daily mortality due to rainfall

fun_sen_loss

Function describing daily mortality due to senescence

fun_dens_fecund

Function describing daily fecundity as function of pest density

fun_temp_fecund

Function describing daily fecundity as function of temperature

fun_age_fecund

Function describing daily fecundity as function of adult cohort age

fun_alate_prod

Function describing proportion of newly produced nymphs destined to become alates as function of pest density

Examples

GPA <-
new("pest",
  species = "Myzus persicae",
  bg_loss = 0.03,
  alate_penalty = 0.5,
  apterae_walk = 0.0005,
  alate_flight = 0.005,
  fun_dev_apt = EndosymbiontModel:::fit_wang(24.1, 0.152, 18.09, 32.79, 2.84, 1.8977),
  fun_dev_ala = EndosymbiontModel:::fit_wang(24.1, 0.16474, 19.96, 33.17, 1.41, 2.2202),
  fun_imi = EndosymbiontModel:::fit_imi_barton(),
  fun_emi = EndosymbiontModel:::fit_emi_thack(),
  fun_temp_loss = EndosymbiontModel:::fit_null(0),
  fun_rainfall_loss = EndosymbiontModel:::fit_rainfall_thack(),
  fun_sen_loss = EndosymbiontModel:::fit_gompertz(0.00188, 0.13),
  fun_dens_fecund = EndosymbiontModel:::fit_bannerman(10000, 0.0008),
  fun_temp_fecund = EndosymbiontModel:::fit_quadratic(-0.7611, 31.9847, 25),
  fun_age_fecund = EndosymbiontModel:::fit_weibull(58, 1.885, 5.953, 0),
  fun_alate_prod = EndosymbiontModel:::fit_alate(67.418, 0.993, 0.076, 300)
 )

GPA
#> An object of class "pest"
#> Slot "species":
#> [1] "Myzus persicae"
#> 
#> Slot "bg_loss":
#> [1] 0.03
#> 
#> Slot "alate_penalty":
#> [1] 0.5
#> 
#> Slot "apterae_walk":
#> [1] 5e-04
#> 
#> Slot "alate_flight":
#> [1] 0.005
#> 
#> Slot "fun_dev_apt":
#> function (x) 
#> {
#>     output <- (H/(1 + exp(-1 * r * (x - Topt)))) * (1 - exp(-1 * 
#>         (x - Tlower)/bound)) * (1 - exp(-1 * (Tupper - x)/bound))/100
#>     output <- ifelse(output < 0, 0, output)
#>     return(output)
#> }
#> <bytecode: 0x5605a2e900c0>
#> <environment: 0x5605a0fe8450>
#> 
#> Slot "fun_dev_ala":
#> function (x) 
#> {
#>     output <- (H/(1 + exp(-1 * r * (x - Topt)))) * (1 - exp(-1 * 
#>         (x - Tlower)/bound)) * (1 - exp(-1 * (Tupper - x)/bound))/100
#>     output <- ifelse(output < 0, 0, output)
#>     return(output)
#> }
#> <bytecode: 0x5605a2e900c0>
#> <environment: 0x5605a0fe7f10>
#> 
#> Slot "fun_imi":
#> function (x) 
#> {
#>     return(5)
#> }
#> <bytecode: 0x5605a33b6228>
#> <environment: 0x5605a0fedb10>
#> 
#> Slot "fun_emi":
#> function (x) 
#> {
#>     output <- round(0.2 * x, 0)
#>     return(output)
#> }
#> <bytecode: 0x5605a45f4b00>
#> <environment: 0x5605a0fed9c0>
#> 
#> Slot "fun_temp_loss":
#> function (x) 
#> {
#>     return(rep(output, length(x)))
#> }
#> <bytecode: 0x56059f95b1f0>
#> <environment: 0x5605a0fed800>
#> 
#> Slot "fun_rainfall_loss":
#> function (x) 
#> {
#>     output <- (0.026 * x) - 0.33
#>     output <- ifelse(output < 0, 0, ifelse(output > 0.5, 0.5, 
#>         output))
#>     return(output)
#> }
#> <bytecode: 0x5605a06a4660>
#> <environment: 0x5605a0fed6e8>
#> 
#> Slot "fun_sen_loss":
#> function (x) 
#> {
#>     output <- (delta * exp(beta * x)) * exp(-1 * (delta/beta) * 
#>         (exp(beta * x) - 1))
#>     output <- ifelse(output > 1, 1, output)
#>     return(output)
#> }
#> <bytecode: 0x5605a0af3ed0>
#> <environment: 0x5605a0fed448>
#> 
#> Slot "fun_dens_fecund":
#> function (x) 
#> {
#>     output <- 1 - (1/(1 + exp(-1 * p * (x - D))))
#>     return(output)
#> }
#> <bytecode: 0x5605a4c8eaa0>
#> <environment: 0x5605a0fed1a8>
#> 
#> Slot "fun_temp_fecund":
#> function (x) 
#> {
#>     output <- (a * x^2 + b * x)/(a * topt^2 + b * topt)
#>     output <- ifelse(output < 0, 0, output)
#>     return(output)
#> }
#> <bytecode: 0x5605a3cd9c78>
#> <environment: 0x5605a0fece98>
#> 
#> Slot "fun_age_fecund":
#> function (x) 
#> {
#>     output <- K * (iota/gamma) * (((x - eta)/gamma)^(iota - 1)) * 
#>         exp(-1 * (((x - eta)/gamma))^iota)
#>     output <- ifelse(is.na(output), 0, output)
#>     return(output)
#> }
#> <bytecode: 0x560594cfcb50>
#> <environment: 0x5605a0fecae0>
#> 
#> Slot "fun_alate_prod":
#> function (x) 
#> {
#>     output <- z/(1 + exp(-1 * W * ((x/V) - omega)))
#>     output <- ifelse(output > 0.5, 0.5, output)
#>     return(output)
#> }
#> <bytecode: 0x5605a1775670>
#> <environment: 0x5605a0fec6f0>
#>