Plant infections

Calculate number of infected plants per time step

Usage

plant_infect(Ph, Pi, Awt, Ai, Awtp, Aip, Ac, temperature, prop_day, Amt)

Arguments

Ph

Population of healthy plants (Plants/m2)

Pi

Population of infected plants (Plants/m2)

Awt

Density of wild type insects (Insects/m2)

Ai

Density of infected insects (Insects/m2)

Awtp

Number of newly produced wild type insects in time step (Insects/m2)

Aip

Number of newly produced infected insects in time step (Insects/m2)

Ac

Carrying capacity of paddock (Insects/m2)

temperature

Mean temperature during time step (C)

prop_day

Proportion of day within temperature thresholds

Amt

Movements of insects per day

Value

Pi the new population of infected plants at time step

Details

Calculate the new population of infected plants at the end of the time step, using the following set of equations: $$ A_{f} = \frac{{A_{wt} + A_{i}}}{{A_{c}}} $$

$$ P_{n} = \frac{{A_{mt} \cdot A_{i}}}{{(A_{wt} + A_{i}) \cdot A_{n}}} $$

$$ P_{i} = P_{h} \cdot \left(1 - e^{-\frac{{P_{n}}}{{P_{h} + P_{i}}}}\right) $$